ra_evaluation_nuhash/Part_B/requirements.txt
keras==2.9.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.2.0
scikit_image==0.19.2
scikit_learn==1.1.1
skimage==0.0
tensorflow==2.9.1
